Delay loading `Parser::Current`
Currently, when the RBI gem is loaded, we get a warning message output 
in the console about a Ruby version mismatch, if the teeny version of 
the running Ruby doesn't match the teeny version of the Parser::Current 
Ruby.

We get the warning message very early since when we require unparser it
requires parser/current immediately. However, the program might not even 
be performing any parsing operations during its execution, so this can 
easily be delayed to parsing time, by doing the require delayed. This
allows executables that use this gem, like Tapioca, to suppress the
warning.