Don't compare default values when comparing `OptParam` and `KwOptParam` nodes
Since the static type checker is not interested in runtime values, but
only in the types of the parameters, RBI files should not care if the default values of parameters are identical or not. Thus, the RBI gem should not care about default values and just use the one from the left-hand side, if they are different.